@@ -3,18 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@dataesr/react-dsfr": "^1.1.8",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.26.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-query": "^3.34.17",
+    "@dataesr/react-dsfr": "^1.1.11",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "axios": "^0.27.2",
+    "copy-to-clipboard": "^3.3.1",
+    "query-string": "^7.1.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-query": "^3.38.1",
+    "react-range": "^1.8.12",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
+    "react-share": "^4.4.0",
+    "react-switch": "^6.0.0",
+    "react-tooltip": "^4.2.21",
+    "recharts": "^2.1.9",
     "sib-api-v3-sdk": "^8.3.0",
     "styled-components": "^5.3.5",
+    "styled-normalize": "^8.0.7",
+    "twemoji": "^14.0.2",
+    "use-query-params": "^1.2.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
